Simple Linear Regression
This commit has a simple linear regression model built to predict height of different mapid (i.e., OpenStreetMap ID for each road segment or edge in the graph). I have chosen regression model to address the problem. The idea is to build a baseline model and see how far can I keep improving to get the predicted height of mapids closer to the ground truth. 

I grouped-by and averaged the features according to each mapid. The result of such a simple linear model is also illustrated using matplotlib.